Rename the namespace `Cycle\ORM\Promise` to `Cycle\ORM\Reference`
