+++ src/Reference/ReferenceInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Cycle\ORM\Promise;
+namespace Cycle\ORM\Reference;
 
 /**
  * Reference points to a remote entity.
