+++ src/Relation/BelongsTo.php
@@ -6,7 +6,7 @@ namespace Cycle\ORM\Relation;
 
 use Cycle\ORM\Exception\Relation\NullException;
 use Cycle\ORM\Heap\Node;
-use Cycle\ORM\Promise\ReferenceInterface;
+use Cycle\ORM\Reference\ReferenceInterface;
 use Cycle\ORM\Relation\Traits\PromiseOneTrait;
 use Cycle\ORM\Transaction\Pool;
 use Cycle\ORM\Transaction\Tuple;
