+++ src/Relation/Morphed/BelongsToMorphed.php
@@ -7,9 +7,9 @@ namespace Cycle\ORM\Relation\Morphed;
 use Cycle\ORM\Exception\RelationException;
 use Cycle\ORM\Heap\Node;
 use Cycle\ORM\ORMInterface;
-use Cycle\ORM\Promise\DeferredReference;
-use Cycle\ORM\Promise\Reference;
-use Cycle\ORM\Promise\ReferenceInterface;
+use Cycle\ORM\Reference\DeferredReference;
+use Cycle\ORM\Reference\Reference;
+use Cycle\ORM\Reference\ReferenceInterface;
 use Cycle\ORM\Relation;
 use Cycle\ORM\Relation\BelongsTo;
 use Cycle\ORM\Transaction\Pool;
